[534] Rename students ui module to presentation
