+++ app/build.gradle.kts
@@ -100,7 +100,7 @@ dependencies {
 
     implementation(project(":feature:syllabus:syllabus-ui"))
 
-    implementation(project(":feature:studentsv2:students-ui"))
+    implementation(project(":feature:studentsv2:presentation"))
 
     implementation(project(":feature:students"))
     implementation(project(":feature:department"))
