+++ feature/studentsv2/presentation/src/main/kotlin/com/elmepa/students/presentation/acadschedule/AcademicScheduleView.kt
@@ -1,4 +1,4 @@
-package com.elmepa.students.ui.acadschedule
+package com.elmepa.students.presentation.acadschedule
 
 import com.stathis.model.UiModel
